+++ R/rbnpsd.R
@@ -26,7 +26,8 @@
 #' # define population structure
 #' F <- c(0.1, 0.3) # FST values for k = 2 subpopulations
 #' sigma <- 1 # dispersion parameter of intermediate subpops
-#' Q <- q1d(n, k, sigma) # admixture proportions from 1D geography
+#' # admixture proportions from 1D geography
+#' Q <- admix_prop_1d_linear(n, k, sigma)
 #'
 #' # draw all random allele freqs and genotypes
 #' out <- rbnpsd(Q, F, m)