+++ R/rgeno.R
@@ -25,7 +25,7 @@
 #' # define population structure
 #' F <- c(0.1, 0.3) # FST values for k=2 subpops
 #' sigma <- 1 # dispersion parameter of intermediate subpops
-#' Q <- q1d(n, k, sigma) # non-trivial admixture proportions
+#' Q <- admix_prop_1d_linear(n, k, sigma) # non-trivial admixture proportions
 #'
 #' # draw alelle frequencies
 #' pAnc <- rpanc(m) # random vector of ancestral allele frequencies